RF: Move basic functionality to SpecialRemote
`AnnexCustomRemote` is the base class for the `datalad` and
`datalad-archives` special remotes only, while `SpecialRemote` is the
more general base class. Hence, move code that is not specific to those
special remotes.
